Updated AI file
Adding updated file

Includes parser for correct SQL queries and some optimizations. 

Changes on the prompt to prevent ambiguity. Including some conditions and deleting unnecesary tables on prompt schemas.

Missing things to possibly do:

Create an additional step in the parser, that checks if the query gives an error when executing the same on the db. Currently, the parser only makes sure that the AI output is in correct SQL format.

Check dates. It seems that the model has issues when we try to look for information on specific date. This should be done by both considering the above point, and maybe changing the source (on tables format is YYYY-MM-DD "XXXXXXXX" (Time), and maybe the specific time (altough is 0 on all cases), is causing issues (I think that probably yes since date filters are placed as strings on most columns of database and SQL.